- Merge the auto-test branch into trunk.
Now the target "ant run-tests" should pass 100% of tests (and prints
a summary per test to let you know it's actually working).

And "ant run-tests-systemtest" should use the red5-selftest.swf
in test/fixtures to make sure things seems to work from FP9.

If you want to add to red5-selftest.swf, the code for that
is in svn/flash/trunk/selftest.  If you update, you'll need
to check in a new SWF to trunk to use that by default.



git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@3254 1b6495e4-3631-0410-8e05-8f51eee8b9cc