Fixed bug http://red5.org/ticket/622. When user issues NetStream.play("streamname"), the default behavior should be the following:
In code, the following is the same:
inStream.play("streamName");
inStream.play("streamName", -2);

On the server we should check the following conditions.

condition #1: look for existing live stream
condition #2: look for existing vod file
condition #3: wait for live stream

if any condition is true, we run the appropriate logic. Previous to this fix, a StreamNotFound would be sent to the client. The conditions used to be:

condition #1: look for existing live stream
condition #2: look for existing vod file
condition #3: send a "StreamNotFound"

git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@4149 1b6495e4-3631-0410-8e05-8f51eee8b9cc